init commit refactor
